Move myqsl tests from test_operators to test_mysql
