+++ tests/operators/test_operators.py
@@ -50,33 +50,6 @@ class TestMySql(unittest.TestCase):
             for table in drop_tables:
                 conn.execute("DROP TABLE IF EXISTS {}".format(table))
 
-    def test_mysql_operator_test(self):
-        sql = """
-        CREATE TABLE IF NOT EXISTS test_airflow (
-            dummy VARCHAR(50)
-        );
-        """
-        from airflow.providers.mysql.operators.mysql import MySqlOperator
-        op = MySqlOperator(
-            task_id='basic_mysql',
-            sql=sql,
-            dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_mysql_operator_test_multi(self):
-        sql = [
-            "CREATE TABLE IF NOT EXISTS test_airflow (dummy VARCHAR(50))",
-            "TRUNCATE TABLE test_airflow",
-            "INSERT INTO test_airflow VALUES ('X')",
-        ]
-        from airflow.providers.mysql.operators.mysql import MySqlOperator
-        op = MySqlOperator(
-            task_id='mysql_operator_test_multi',
-            sql=sql,
-            dag=self.dag,
-        )
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
     def test_mysql_hook_test_bulk_load(self):
         records = ("foo", "bar", "baz")
 
@@ -145,27 +118,6 @@ class TestMySql(unittest.TestCase):
             sql=sql,
             dag=self.dag)
         op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_overwrite_schema(self):
-        """
-        Verifies option to overwrite connection schema
-        """
-        from airflow.providers.mysql.operators.mysql import MySqlOperator
-
-        sql = "SELECT 1;"
-        op = MySqlOperator(
-            task_id='test_mysql_operator_test_schema_overwrite',
-            sql=sql,
-            dag=self.dag,
-            database="foobar",
-        )
-
-        from _mysql_exceptions import OperationalError
-        try:
-            op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE,
-                   ignore_ti_state=True)
-        except OperationalError as e:
-            assert "Unknown database 'foobar'" in str(e)
 
 
 @pytest.mark.backend("postgres")
